requirements.txt
ultralytics==8.0.20
opencv-python==4.10.0.82
python-docx==1.1.0
Pillow==9.5.0
